add nuevos roles(admin- jefe museo - jefe personal- personal administrativo) y filtrado de todos los botones de acuerdo a estos roles, export db, variable global conf.params.idAdmin->1 se usa para controlar que no se borre o actualice el user admin,  Closes #13,#18,#19

+++ phpcode/controllers/UserController.php
@@ -15,9 +15,9 @@ use kartik\mpdf\Pdf;
 
 
 
-class UserController extends MainController
+class UserController extends MainController 
 {
-    
+    // private $idAdmin=1;
     public function behaviors()
     {
         return array_merge(
@@ -40,8 +40,9 @@ class UserController extends MainController
     public function actionIndex()
     {
         $searchModel = new UserSearch();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
+        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+       // var_dump(Yii::$app->request->queryParams);die();
         return $this->render('index', [
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
@@ -55,9 +56,12 @@ class UserController extends MainController
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        if($id!=Yii::$app->params['idAdmin']){
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+         }
+         return $this->actionIndex();
     }
 
     /**
@@ -92,15 +96,32 @@ class UserController extends MainController
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
-        } else {
-            return $this->render('update', [
-                'model' => $model,
-            ]);
-        }
+        if($id==Yii::$app->params['idAdmin'])
+            {
+                         Yii::$app->session->setFlash('error',
+                        [
+                            //'type' => 'error',
+                            'icon' => 'fa fa-users',
+                            'message' => ' no esta permitido.', 
+                            'title' => 'Error de Actualización',
+                            'positonY' => 'top',
+                            'positonX' => 'left'
+                        ]                    
+                );
+                  return $this->redirect(['view','id'=>$id]);
+            }else{
+
+                    $model = $this->findModel($id);
+
+                    if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                        return $this->redirect(['view', 'id' => $model->id]);
+                    } else {
+                        return $this->render('update', [
+                            'model' => $model,
+                        ]);
+                    }
+                }
     }
 
     /**
@@ -111,16 +132,32 @@ class UserController extends MainController
      */
     public function actionDelete($id)
     {
+
         try {
-            if($this->findModel($id)->delete()){
-                return $this->redirect(['index']);
+            if($id!=Yii::$app->params['idAdmin']){
+                if($this->findModel($id)->delete()){
+                    return $this->redirect(['index']);
+                }
+            }else{
+                 Yii::$app->session->setFlash('error',
+                [
+                    //'type' => 'error',
+                    'icon' => 'fa fa-users',
+                    'message' => ' no esta permitido.', 
+                    'title' => 'Error de Borrado',
+                    'positonY' => 'top',
+                    'positonX' => 'left'
+                ]                    
+            );
+                  return $this->redirect(['view','id'=>$id]);
             }
+
         } catch (\yii\db\IntegrityException $exc) {
             Yii::$app->session->setFlash('error',
                 [
                     //'type' => 'error',
                     'icon' => 'fa fa-users',
-                    'message' => 'Usuario posee elementos relacionados',
+                    'message' => 'Usuario posee elementos relacionados, o no esta permitido.', 
                     'title' => 'Error de Borrado',
                     'positonY' => 'top',
                     'positonX' => 'left'
